Add files via upload
Note from email to Andy Feb 11 2020

I got a little lost in all the versions that I have created, so I walked through the process again to ensure it is correct for both the wetlands dataset and the lakes. I believe December 9 version had an upper limit on the size, which I removed this time around. I don't think this dataset adds many wetlands, but it didn't make sense to me to have any size restrictions at this point. It should be easy to restrict sizes if you choose to do so. 

In the end I re-ran it on lakes and all wetlands regardless of size. The original shapefiles that I used is also included. The datasets should link to each other through a unique id if you want to bring the csv files in to ArcGIS.